requirements.txt
asyncpg==0.21.0
databases==0.4.1
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.15.0
rsa==4.7
SQLAlchemy==1.3.22
Telethon==1.19.1
alembic~=1.5.2
telebot~=0.0.4